+++ backend/src/helpers/authentication-helpers.ts
