--- backend/src/models/doctorModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import doctorSchema from '../db/doctorSchema';
-
-const DoctorModel = mongoose.model('Doctor', doctorSchema);
-
-export default DoctorModel;
