--- backend/src/models/queueModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import queue from '../db/queueSchema';
-
-const QueueModel = mongoose.model('Queue', queueSchema);
-
-export default QueueModel;
