--- backend/src/models/specializationModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import specialization from '../db/specializationSchema';
-
-const SpecializationModel = mongoose.model('Specialization', specializationSchema);
-
-export default SpecializationModel
