--- backend/src/models/userInQueueModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import userInQueue from '../db/userInQueue';
-
-const UserInQueueModel = mongoose.model('UserInQueue', userInQueueSchema);
-
-export default UserInQueueModel
