--- backend/src/models/userModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import userSchema from '../db/userSchema';
-
-const UserModel = mongoose.model('User', userSchema);
-
-export default UserModel;
