--- backend/src/models/visitCategoryModel.ts
@@ -1,6 +1,0 @@
-import mongoose from 'mongoose';
-import visitCategorySchema from '../db/visitCategorySchema';
-
-const VisitCategoryModel = mongoose.model('visitCategory', visitCategorySchema);
-
-export default VisitCategoryModel;
